refactor(standalone): create a different package
